+++ components/basic_widgets/SmallDropDown.vue
@@ -38,7 +38,7 @@
 </template>
 <script>
 export default {
-  name: "NewSelectDropDown",
+  name: "SmallDropDown",
   props: {
     options_text: { type: Array, required: true },
     input_width: { type: Number, default: 210 },
@@ -81,6 +81,15 @@ export default {
   created() {
     this.chosen_option = this.dropdown_options[0];
     this.options_list = this.dropdown_options;
+    this.unsubscribe = this.$store.subscribe(async (mutation) => {
+      if (mutation.type === "stimulation/reset_state") {
+        this.chosen_option = this.dropdown_options[0];
+        this.changeSelection(0);
+      }
+    });
+  },
+  beforeDestroy() {
+    this.unsubscribe();
   },
   methods: {
     changeSelection(idx) {