+++ tests/unit/store/stimulation.spec.js
@@ -121,9 +121,8 @@ describe("store/stimulation", () => {
     });
 
     test("When a user requests to delete the current stimulation by using the trash icon, Then it should mutate state to true", async () => {
-      expect(store.state.stimulation.delete_protocol).toBe(false);
-      await store.commit("stimulation/handle_delete_protocol");
-      expect(store.state.stimulation.delete_protocol).toBe(true);
+      await store.commit("stimulation/reset_state");
+      expect(store.state.stimulation.x_axis_points).toStrictEqual([]);
     });
 
     test("When a user selects a new stimulation type to Current Stimulation Type, Then it should mutate state Current", async () => {